Fix ROI int/float handling in motion analyses
ROIs are silently assumed to be 4 interger numbers (x,y,w,h) but
the tracking_rois option changes them and produces non-interger
ROIs.

Previously in 3 out of 4 places the ROI was changed into ingeters
before calling a cc backend (and in one place accidentally not
causing a bug that tkgui with tracking_rois won't work).

Now the ROI is sent to the backend as a interger or floating point
numer and it is the job of the backend to descide if it wants
to use the extra information of the floating point or just
integerify the ROI.

Our opencv backend currently just integerifies